requirements.txt
fastapi>=0.110.0
uvicorn>=0.30.0
openai>=1.30.0
anthropic>=0.50.0
python-dotenv>=1.0.0
